Make state a recycler wrapper, discard extra bits.
